+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+### Changed
+- Renamed `ReceivedRequest::next_message()` to `recv_update()`.
+- Split `SentRequest::next_message()` in `recv_update()` and `recv_response()`.
+
 ## v0.3.1 - 2021-01-27
 ### Changed
 - Update to tokio-seqpacket 0.5.